# at_3/api/openapi.yaml
openapi: 3.0.3
info:
  title: F-tool API
  version: 2.0.0
servers:
  - url: /api/v1
...
paths:
...
  /resource-auth/{resourceType}/{resourceId}:
    get:
      tags: [tables]
      operationId: listResourceAuth
      summary: リソースのアクセス許可一覧 (settings:admin)
      parameters:
        - $ref: '#/components/parameters/ResourceType'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                required: [entries]
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResourceAuthEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /resource-auth/{resourceType}/{resourceId}/{subjectKind}/{subjectId}:
    put:
      tags: [tables]
      operationId: setResourceAuth
      summary: リソースのアクセス許可を付与/変更 (settings:admin)
      parameters:
        - $ref: '#/components/parameters/ResourceType'
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/SubjectKind'
        - $ref: '#/components/parameters/SubjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceAuthSet'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceAuthEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [tables]
      operationId: deleteResourceAuth
      summary: リソースのアクセス許可を削除 (settings:admin)
      parameters:
        - $ref: '#/components/parameters/ResourceType'
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/SubjectKind'
        - $ref: '#/components/parameters/SubjectId'
      responses:
        '204':
          description: 削除完了
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
  parameters:
...
    ResourceType:
      name: resourceType
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ResourceType'
    ResourceId:
      name: resourceId
      in: path
      required: true
      schema:
        type: integer
      example: 1
    SubjectKind:
      name: subjectKind
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/SubjectKind'
    SubjectId:
      name: subjectId
      in: path
      required: true
      schema:
        type: string
      example: df714f28-1a27-1041-8c21-ad329d0f95ce
  responses:
    Unauthorized:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - invalid_credentials
            - unauthorized
            - forbidden
            - not_found
            - validation_failed
            - conflict
            - ldap_unavailable
            - session_unavailable
            - connection_unavailable
            - internal
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    ResourceAuthLevel:
      type: string
      enum: [r, rw]
    ResourceType:
      type: string
      enum: [action, managed_table]
    SubjectKind:
      type: string
      enum: [user, group]
    ResourceAuthEntry:
      type: object
      required: [subjectKind, subjectId, name, authLevel]
      properties:
        subjectKind:
          $ref: '#/components/schemas/SubjectKind'
        subjectId:
          type: string
        name:
          type: string
        username:
          type: string
        authLevel:
          $ref: '#/components/schemas/ResourceAuthLevel'
    ResourceAuthSet:
      type: object
      required: [authLevel]
      properties:
        authLevel:
          $ref: '#/components/schemas/ResourceAuthLevel'
...
security:
  - cookieAuth: []